Use direct CDN url for assets when run on glitch when possible
Previously when run as a glitch project any requests for assets were
transparently proxied to their CDN url by the backend. This is still the
case however when possible we now try to load asset directly from CDN
url. This makes the client end code a bit more complex however glitch
limits the number of requests that can be made to project domain over a
certain amount of time so it's important to keep this as low as
possible.